+++ irisweb/controller.go
@@ -1,0 +1,13 @@
+package irisweb
+
+import (
+	"github.com/kataras/iris/v12"
+)
+
+// Controller controller
+type Controller interface {
+	AddHandle(handle iris.Handler)
+	ClearHandle()
+	GetHandles() []iris.Handler
+	GetRoute() Route
+}