+++ irisweb/irisweb.type.go
@@ -4,7 +4,7 @@ import (
 	"github.com/kataras/iris/v12"
 )
 
-// Configure Web configure
+// Configure web configure
 type Configure struct {
 	Host string             `json:"host"`
 	Port int                `json:"port"`
@@ -36,10 +36,11 @@ type Paths struct {
 
 // Route Web route
 type Route struct {
-	Name   string
-	Path   string
-	Method string
-	Handle []iris.Handler
+	Handle  []iris.Handler
+	Method  string
+	Name    string
+	Path    string
+	Summary string
 }
 
 // StaticFile Web static file
@@ -54,22 +55,4 @@ type StaticFile struct {
 type StaticPath struct {
 	Absolute string `json:"absolute"`
 	Relative string `json:"relative"`
-}
-
-// Web Web service
-type Web interface {
-	CatchHandleError(ctx iris.Context, handleName string)
-	GetIris() *iris.Application
-	GetPaths() Paths
-	GetStaticFileString(pathname, filename string) string
-	GetStaticPaths(name string) (StaticPath, error)
-	RemoveFileFromStaticFileString(fileString string) error
-	SetRoutes(routes []Route) error
-	SetStaticPaths(name string, path StaticPath, options iris.DirOptions) error
-	StaticFileStringToFile(fileString string) (StaticFile, error)
-	StaticFileToString(file StaticFile) string
-	StaticFileToURL(file StaticFile) string
-	ThrowHandleError(statucCode int, params HandleErrorParams)
-	Start() error
-	Stop() error
 }