+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.0.2] - 2020-07-23
+
+### Added
+
+- Add GoLibrary.NewWebController
+
+### Changed
+
+- Adjust Web Route, Add property Summary
+
 ## [v1.0.1] - 2020-07-22
 
 ### Changed
@@ -19,5 +29,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add logger
 - Add miscellaneous
 
+[v1.0.2]: https://github.com/CloudChen2018/golib-definitions/archive/v1.0.2.zip
 [v1.0.1]: https://github.com/CloudChen2018/golib-definitions/archive/v1.0.1.zip
 [v1.0.0]: https://github.com/CloudChen2018/golib-definitions/archive/v1.0.0.zip